requirements.txt
google-generativeai
openai==0.28.1
langchain==0.0.330
discord.py[voice]
python-dotenv
requests
pretty_errors
mistralai
dotenv
youtube-search-python
google-api-python-client
